Revert "Fix 1st param of GetByteArray to ULONG"
This reverts commit b4a48c1d3fc9eeb412c44e0c65e4a420dbea4c73.

+++ com/TestSources/PyCOMTest/PyCOMImpl.cpp
@@ -379,9 +379,9 @@ STDMETHODIMP CPyCOMTest::GetSafeArrays(SAFEARRAY **attrs, SAFEARRAY **attrs2, SA
     return S_OK;
 }
 
-STDMETHODIMP CPyCOMTest::GetByteArray(ULONG sizeBytes, SAFEARRAY **array)
+STDMETHODIMP CPyCOMTest::GetByteArray(long sizeBytes, SAFEARRAY **array)
 {
-    SAFEARRAYBOUND bound = {sizeBytes, 0};
+    SAFEARRAYBOUND bound = {static_cast<ULONG>(sizeBytes), 0};
     *array = SafeArrayCreate(VT_UI1, 1, &bound);
     return S_OK;
 }
